requirements.txt
# AWS SDK
boto3>=1.34.0
botocore>=1.34.0

# AgentCore and Strands
strands-sdk>=0.1.0

# API Clients
PyGithub>=2.1.1
jira>=3.5.0
requests>=2.31.0

# Azure DevOps (via MCP proxy — no extra deps beyond requests)

# Security Analysis
bandit>=1.7.5
safety>=2.3.5

# Code Quality
pylint>=3.0.0
radon>=6.0.1

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0

# Development
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
moto>=4.2.0
black>=23.12.0
mypy>=1.7.0
flake8>=6.1.0
